+++ .gitignore
@@ -132,3 +132,4 @@ dmypy.json
 # data/
 # tdet_run.py
 # treg_run.py
+__process.py
