+++ __init__.py
@@ -1,1 +1,1 @@
-from .mebit.metrics import Evaluation
+from .mebit.__process import Evaluation
