+++ mebit/classification.py
@@ -1,0 +1,19 @@
+from .base import BaseEvaluation
+from .utils.util import HiddenPrints
+
+class ClsfEvaluation(BaseEvaluation):
+    def save_gt(self, gt, img_names):
+        ...
+    def save_dt(self, dt, img_names):
+        ...
+    def evaluate(self, gt, dt):
+        ...
+    def read_groundtruth(self):
+        ...
+    def format_original_gt(self, *args):
+        ...
+    def format_transform_gt(self, *args):
+        ...
+    def format_dt(self):
+        ...
+
