--- mebit/text_recognition.py
@@ -1,2 +1,0 @@
-from .base import BaseEvaluation
-
